Now the database sourced unit test data work. I cleaned up the naming of the datasets and tests.
Next I will need to contribute a global test runner that can run
both CSV and DB backed tests.

The job in the PDI-17034_db_checksum handles orchestration and waiting for
the mariadb container to come fully online with populated data.

Naming the input tables 'input' and output tables 'output' with
a standard schema name 'testdata' really helped.
The database connection for that is shared.

Naming convention is database_testData for the connections.